*refactoring of ConvertReduceSumNode.
